[C API] Refactor C API implementation to make it portable to the public repo
